requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
slack-sdk==3.26.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pymysql==1.1.0
cryptography==41.0.7 